add report.MatplotlibFigure
